+++ changelog.md
@@ -7,6 +7,7 @@
 * Update CSIRO renderer (1.2.11) and template extraction (1.0.14) engines
 * Update CSIRO renderer to use the terminology server URL from the settings page
 * fhirpath tester page share link can now support using bookmarks to support much longer links
+* Update axios dependency to 1.13.2
 
 # 26 December 2025
 * Updated the static config loading code to use async/await to ensure that the settings are fully loaded before being used.